Add files via upload
Number Inpatient

This feature shows how many prior inpatient visits a patient has had. Frequent past hospitalizations are a strong signal of chronic health issues or poor disease management, both of which increase the likelihood of readmission. This field is therefore an excellent indicator of readmission status and should be included in the model.